Move HTTPHeaderDict
