+++ yt_dlp/downloader/fragment.py
@@ -12,13 +12,8 @@ from ..aes import aes_cbc_decrypt_bytes, unpad_pkcs7
 from ..compat import compat_os_name
 from ..networking import Request
 from ..networking.exceptions import HTTPError, IncompleteRead
-from ..utils import (
-    DownloadError,
-    HTTPHeaderDict,
-    RetryManager,
-    encodeFilename,
-    traverse_obj,
-)
+from ..networking.utils import HTTPHeaderDict
+from ..utils import DownloadError, RetryManager, encodeFilename, traverse_obj
 
 
 class HttpQuietDownloader(HttpFD):
