+++ yt_dlp/downloader/youtube_live_chat.py
@@ -3,8 +3,8 @@ import time
 
 from .fragment import FragmentFD
 from ..networking.exceptions import HTTPError
+from ..networking.utils import HTTPHeaderDict
 from ..utils import (
-    HTTPHeaderDict,
     RegexNotFoundError,
     RetryManager,
     dict_get,
