+++ yt_dlp/utils/_utils.py
@@ -5136,32 +5136,6 @@ def truncate_string(s, left, right=0):
     return f'{s[:left-3]}...{s[-right:] if right else ""}'
 
 
-class HTTPHeaderDict(collections.UserDict, dict):
-    """
-    Store and access keys case-insensitively.
-    The constructor can take multiple dicts, in which keys in the latter are prioritised.
-    """
-
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-        for dct in args:
-            if dct is not None:
-                self.update(dct)
-        self.update(kwargs)
-
-    def __setitem__(self, key, value):
-        super().__setitem__(key.title(), value)
-
-    def __getitem__(self, key):
-        return super().__getitem__(key.title())
-
-    def __delitem__(self, key):
-        super().__delitem__(key.title())
-
-    def __contains__(self, key):
-        return super().__contains__(key.title() if isinstance(key, str) else key)
-
-
 def orderedSet_from_options(options, alias_dict, *, use_regex=False, start=None):
     assert 'all' in alias_dict, '"all" alias is required'
     requested = list(start or [])
@@ -5487,38 +5461,6 @@ class FormatSorter:
             format['tbr'] = try_call(lambda: format['vbr'] + format['abr']) or None
 
         return tuple(self._calculate_field_preference(format, field) for field in self._order)
-
-
-def clean_proxies(proxies: dict, headers: HTTPHeaderDict):
-    req_proxy = headers.pop('Ytdl-Request-Proxy', None)
-    if req_proxy:
-        proxies.clear()  # XXX: compat: Ytdl-Request-Proxy takes preference over everything, including NO_PROXY
-        proxies['all'] = req_proxy
-    for proxy_key, proxy_url in proxies.items():
-        if proxy_url == '__noproxy__':
-            proxies[proxy_key] = None
-            continue
-        if proxy_key == 'no':  # special case
-            continue
-        if proxy_url is not None:
-            # Ensure proxies without a scheme are http.
-            proxy_scheme = urllib.request._parse_proxy(proxy_url)[0]
-            if proxy_scheme is None:
-                proxies[proxy_key] = 'http://' + remove_start(proxy_url, '//')
-
-            replace_scheme = {
-                'socks5': 'socks5h',  # compat: socks5 was treated as socks5h
-                'socks': 'socks4'  # compat: non-standard
-            }
-            if proxy_scheme in replace_scheme:
-                proxies[proxy_key] = urllib.parse.urlunparse(
-                    urllib.parse.urlparse(proxy_url)._replace(scheme=replace_scheme[proxy_scheme]))
-
-
-def clean_headers(headers: HTTPHeaderDict):
-    if 'Youtubedl-No-Compression' in headers:  # compat
-        del headers['Youtubedl-No-Compression']
-        headers['Accept-Encoding'] = 'identity'
 
 
 # XXX: Temporary
